requirements.txt
alembic==1.8.1
appdirs==1.4.4
async-generator==1.10
attrs==21.4.0
azure-ai-formrecognizer==3.2.0
azure-common==1.1.28
azure-core==1.26.0
backoff==1.11.1
beautifulsoup4==4.11.1
blis==0.7.8
bs4==0.0.1
camelot-py==0.10.1
catalogue==2.0.8
certifi==2022.6.15
cffi==1.15.0
chardet==5.0.0
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.2.0
cryptography==37.0.2
cssselect==1.1.0
cymem==2.0.6
databricks-cli==0.17.3
dill==0.3.5.1
docker==6.0.0
docopt==0.6.2
elastic-apm==6.12.0
elasticsearch==7.10.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.4.0/en_core_web_lg-3.4.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.4.0/en_core_web_trf-3.4.0-py3-none-any.whl
entrypoints==0.4
et-xmlfile==1.1.0
fake-useragent==0.1.11
farm-haystack==1.9.1
filelock==3.8.0
Flask==2.1.2
Flask-Cors==3.0.10
fuzzywuzzy==0.18.0
gevent==21.12.0
gitdb==4.0.9
GitPython==3.1.29
greenlet==1.1.3
gunicorn==20.1.0
h11==0.13.0
huggingface-hub==0.9.0
idna==3.3
importlib-metadata==4.11.4
inflect==6.0.0
isodate==0.6.1
itsdangerous==2.1.2
jarowinkler==1.2.3
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.16.0
langcodes==3.3.0
langdetect==1.0.9
lxml==4.9.0
Mako==1.2.3
MarkupSafe==2.1.1
mlflow==1.29.0
mmh3==3.0.0
monotonic==1.6
more-itertools==8.14.0
msrest==0.7.1
murmurhash==1.0.7
mysql-connector==2.2.9
networkx==2.8.7
nltk==3.7
num2words==0.5.12
numpy==1.23.0
oauthlib==3.2.1
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.4.3
parse==1.19.0
pathy==0.6.2
pdfminer.six==20220524
pdfminer3k==1.3.2
Pillow==9.2.0
posthog==2.1.2
preshed==3.0.6
prometheus-client==0.15.0
prometheus-flask-exporter==0.20.3
protobuf==4.21.7
pyahocorasick==1.4.4
pycparser==2.21
pydantic==1.9.1
pyee==8.2.2
PyJWT==2.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyPDF2==2.10.9
pyppeteer==1.0.2
pyquery==1.4.3
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.6
pytz==2022.1
PyYAML==6.0
quantulum3==0.7.10
querystring-parser==1.2.4
rapidfuzz==2.7.0
regex==2022.7.25
requests==2.28.0
requests-html==0.10.0
requests-oauthlib==1.3.1
scikit-learn==1.1.2
scipy==1.9.0
selenium==4.3.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
seqeval==1.2.2
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-alignments==0.8.5
spacy-legacy==3.0.9
spacy-loggers==1.0.3
spacy-transformers==1.1.8
SQLAlchemy==1.4.41
sqlparse==0.4.3
srsly==2.4.4
tabulate==0.8.10
thinc==8.1.0
threadpoolctl==3.1.0
tika==1.24
tokenizers==0.12.1
torch==1.10.1
torchtext==0.11.1
torchvision==0.13.1
tqdm==4.64.0
transformers==4.21.2
trio==0.21.0
trio-websocket==0.9.2
typer==0.4.2
typing-extensions==4.2.0
urllib3==1.26.9
w3lib==1.22.0
waitress==2.1.2
wasabi==0.10.1
websocket-client==1.4.1
websockets==10.3
Werkzeug==2.1.2
wsproto==1.1.0
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
